Switch to int16s for multi-target

+++ photon-lib/src/main/native/include/photonlib/MultiTargetPNPResult.h
@@ -52,7 +52,7 @@ class PNPResults {
 class MultiTargetPnpResult {
  public:
   PNPResults result;
-  wpi::SmallVector<int8_t, 32> fiducialIdsUsed;
+  wpi::SmallVector<int16_t, 32> fiducialIdsUsed;
 
   friend Packet& operator<<(Packet& packet, const MultiTargetPnpResult& result);
   friend Packet& operator>>(Packet& packet, MultiTargetPnpResult& result);
